QUANT-60 테스트용 컨트롤러 구현

+++ src/main/java/stock/chart/stock/dto/StockPriceDto.java
@@ -18,6 +18,6 @@ public class StockPriceDto {
     private int high;
     private int low;
     private int close;
-    private Long volume;
+    private int volume;
 
 }
